refactor: move DynamicModelInfo into new file to eliminate duplication and circular imports
